Fix hwaddr setting and add persistence
MAC addresses have a 3-byte Organisationally Unique Identifier and a
3-byte NIC number. Bit 1 of the first OUI octet should be '1' to signify
a locally administered address, and bit '0' MUST be '0' to signify a
unicast packet. Use of locally administered addresses avoids collision
with IEEE assigned addresses.

This commit adds persistence to the OUI part of the MAC address. The
first time it's run, a random 3-octet address is chosen to comply with
the above rules, and stored in the config file. Subsequent calls
reuse the same 3-octet prefix.

The NIC part of the MAC address comes from the SHA1 hash of the
container name.

This provides 24 bits of randomness (16 million unique hosts) and
24 bits of similarity (the first 3 octets of the MAC address).

Signed-off-by: Nick Andrew <nick@nick-andrew.net>